requirements.txt
allure-pytest==2.8.6
pandas==0.25.3
paramiko==2.6.0
pytest==5.3.2
pytest-html==2.0.1
pyyaml==5.1.2
robotframework==3.1.2
robotframework-selenium2library==3.0.0
selenium==3.141.0
wheel==0.33.6